requirements.txt
absl-py==2.1.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
fastjsonschema==2.20.0
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.19.1
jinja2==3.1.4
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
keras==2.10.0
keras-preprocessing==1.1.2
libclang==18.1.1
markdown==3.4.4
markupsafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
notebook==6.5.7
notebook-shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opt-einsum==3.3.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==9.5.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygments==2.17.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pywin32==306
pywinpty==2.0.10
pyzmq==26.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.7.3
send2trash==1.8.3
setuptools==65.6.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
split-folders==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
